fix 管理者ではないときの処理追加

+++ api/app/Http/Controllers/Admin/CircleUser/DeleteCircleUserController.php
@@ -7,7 +7,9 @@ namespace App\Http\Controllers\Admin\CircleUser;
 use App\Http\Controllers\Controller;
 use App\Models\Circle;
 use App\Models\User;
+use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 final class DeleteCircleUserController extends Controller
 {
@@ -15,8 +17,23 @@ final class DeleteCircleUserController extends Controller
     {
         Circle::findOrFail($circleId);
         $user = User::findOrFail($userId);
-        $user->circleUsers()->delete();
-        $user->delete();
+
+        DB::beginTransaction();
+        try {
+            // サークル管理者の削除
+            $user->circleUsers()->delete();
+
+            // 管理者ではないとき
+            if (!$user->isAdminUser()) {
+                $user->delete();
+            }
+
+            DB::commit();
+        } catch (Exception $e) {
+            DB::rollBack();
+
+            throw $e;
+        }
 
         return [
             'status' => true,
